refactor(logger): remove BaseLogger
