+++ src/logger/impl.ts
@@ -1,12 +1,33 @@
-import { Injectable } from '@artus/injection';
-import { BaseLogger } from './base';
-import { LoggerLevel } from './level';
-import { LogOptions } from './types';
+import { Container, Inject, Injectable, ScopeEnum } from '@artus/injection';
+import { ArtusInjectEnum } from '../constant';
+import { LoggerLevel, LOGGER_LEVEL_MAP } from './level';
+import { LoggerOptions, LoggerType, LogOptions } from './types';
 
 @Injectable({
-  scopeEscape: true
+  scope: ScopeEnum.SINGLETON
 })
-export default class Logger extends BaseLogger {
+export default class Logger implements LoggerType {
+  @Inject()
+  protected container!: Container;
+
+  protected get loggerOpts(): LoggerOptions {
+    let appConfig: Record<string, any> = {};
+    try {
+      appConfig = this.container.get(ArtusInjectEnum.Config);
+    } catch(e) {
+      // do nothing
+    }
+    return appConfig?.logger ?? {};
+  }
+
+  protected checkLoggerLevel(level: LoggerLevel) {
+    const targetLevel = this.loggerOpts.level ?? LoggerLevel.INFO;
+    if (LOGGER_LEVEL_MAP[level] < LOGGER_LEVEL_MAP[targetLevel]) {
+      return false;
+    }
+    return true;
+  }
+
   public trace(message: string, ...args: any[]) {
     if (!this.checkLoggerLevel(LoggerLevel.TRACE)) {
       return;
